tests: add some basic functionality tests
